cli/requirements.txt
click
rich
httpx
pyyaml